Extract 10 pure value-coercion primitives into modules/output_values.py
First refactor pass at modules/output.py (3833 lines, well over the
600-line soft cap).  Moved ten small, dependency-free parsing helpers
that only touch stdlib (json / ast / re):

  * _normalize_template_value  -- unwrap {"value": X} and trim
  * _coerce_bool               -- bool | 'true'/'false' string
  * _to_number                 -- int/float from string
  * _coerce_string_list        -- dedup + strip + drop empties/brackets
  * _parse_string_list         -- list-shaped input normalizer
  * _parse_comma_string_list   -- CSV or JSON/py-literal list
  * _parse_string_list_mapping -- {key: [values]} normalizer
  * _parse_string_mapping      -- {key: 'joined, values'} normalizer
  * _parse_template_mapping_dict -- generic dict-from-anything
  * _playlist_scalar_or_list   -- collapse 1-elem list to scalar

These are pure functions with no Flask / helpers / persistence /
database dependencies and were used dozens of times across output.py.
output.py now imports them explicitly at the top so historical call
sites like output._parse_string_list(...) (used by tests) still work.

Impact:
- modules/output.py: 3833 -> 3670 lines (-163 / -4.3%)
- modules/output_values.py: 203 lines new
- Net +40 lines total (docstring + explicit import list) in exchange
  for a testable, dependency-free primitive module
- Zero behavior change; every symbol is byte-identical to its former
  in-file definition

All 1077 tests pass. Ruff + black clean.
